+++ Project/breathing_led.h
@@ -13,8 +13,8 @@
 #include "gpio_port.h"
 #include "main.h"
 
-extern uint8_t R, G, B;
-extern uint8_t R_temp, G_temp, B_temp;
+extern uint16_t R, G, B;
+extern uint16_t R_temp, G_temp, B_temp;
 extern bool RGB_update, RGB_direction;
 
 void breathing_led_init(void);
